REMOVE THIS COMMIT BEFORE MERGING.
Rename `get_option` to `get_option2` so all uses of it in the
code base create an error.

In the next commit add a common function for getting option
values which just calls `get_option2`.

In the last one make it call `get_option`.

When everything is done, squash all three into one.